+++ wiki/ReleaseNotes.md
@@ -6,6 +6,7 @@ none
 ## Other changes
 * **[JEE]** Applications can inject a MqttClientIdProvider to generate an id for the mqtt client.
   The producer method must be annotated with JoynrMqttClientIdProvider.
+* **[Java,C++,JS]** updated SMRF dependency to 0.2.0 which introduces an incompatibility with any previous version
 
 # joynr 0.27.1