+++ wiki/ReleaseNotes.md
@@ -7,6 +7,7 @@ none
 * **[JEE]** Applications can inject a MqttClientIdProvider to generate an id for the mqtt client.
   The producer method must be annotated with JoynrMqttClientIdProvider.
 * **[Java,C++,JS]** updated SMRF dependency to 0.2.0 which introduces an incompatibility with any previous version
+* **[C++]** Fixed potential crash when a proxy is used after a joynr runtime was deleted.
 
 # joynr 0.27.1